Add JIT `execute` function with an easy-ish to use API.
