+++ Cargo.toml
@@ -12,7 +12,7 @@ name = "sierrajit"
 required-features = ["build-cli"]
 
 [features]
-build-cli = ["clap", "tracing-subscriber"]
+build-cli = ["clap", "serde_json", "tracing-subscriber"]
 
 [dependencies]
 bumpalo = "3.13.0"
@@ -29,7 +29,7 @@ lazy_static = "1.4.0"
 libc = "0.2.146"
 melior = "0.9.4"
 mlir-sys = "0.2.0"
-num-bigint = "0.4.3"
+num-bigint = { version = "0.4.3", features = ["serde"] }
 serde = { version = "1.0.164", features = ["derive"] }
 tracing = "0.1.37"
 
@@ -39,6 +39,9 @@ serde_json = { version = "1.0.97", optional = true }
 tracing-subscriber = { version = "0.3.17", features = [
     "env-filter",
 ], optional = true }
+
+[dev-dependencies]
+serde_json = "1.0.97"
 
 [build-dependencies]
 cc = "1.0.79"
