+++ src/values/felt252.rs
@@ -1,11 +1,51 @@
 use super::ValueBuilder;
+use crate::types::felt252::PRIME;
 use cairo_lang_sierra::{
     extensions::{types::InfoOnlyConcreteType, GenericLibfunc, GenericType},
     ids::ConcreteTypeId,
     program_registry::ProgramRegistry,
 };
 use num_bigint::BigUint;
+use serde::{Deserialize, Deserializer, Serialize, Serializer};
 use std::{fmt, ptr::NonNull};
+
+pub unsafe fn deserialize<'de, TType, TLibfunc, D>(
+    deserializer: D,
+    _registry: &ProgramRegistry<TType, TLibfunc>,
+    ptr: NonNull<()>,
+    _info: &InfoOnlyConcreteType,
+) -> Result<(), D::Error>
+where
+    TType: GenericType,
+    TLibfunc: GenericLibfunc,
+    <TType as GenericType>::Concrete: ValueBuilder<TType, TLibfunc>,
+    D: Deserializer<'de>,
+{
+    let value = <BigUint as Deserialize>::deserialize(deserializer)?;
+    assert!(value < *PRIME);
+
+    let mut bytes = value.to_bytes_le();
+    bytes.resize(32, 0);
+    ptr.cast::<[u8; 32]>().as_mut().copy_from_slice(&bytes);
+
+    Ok(())
+}
+
+pub unsafe fn serialize<TType, TLibfunc, S>(
+    serializer: S,
+    _registry: &ProgramRegistry<TType, TLibfunc>,
+    ptr: NonNull<()>,
+    _info: &InfoOnlyConcreteType,
+) -> Result<S::Ok, S::Error>
+where
+    TType: GenericType,
+    TLibfunc: GenericLibfunc,
+    <TType as GenericType>::Concrete: ValueBuilder<TType, TLibfunc>,
+    S: Serializer,
+{
+    let value = BigUint::from_bytes_le(ptr.cast::<[u8; 32]>().as_ref());
+    <BigUint as Serialize>::serialize(&value, serializer)
+}
 
 pub unsafe fn debug_fmt<TType, TLibfunc>(
     f: &mut fmt::Formatter,
@@ -17,7 +57,7 @@ pub unsafe fn debug_fmt<TType, TLibfunc>(
 where
     TType: GenericType,
     TLibfunc: GenericLibfunc,
-    <TType as GenericType>::Concrete: ValueBuilder,
+    <TType as GenericType>::Concrete: ValueBuilder<TType, TLibfunc>,
 {
     fmt::Debug::fmt(&BigUint::from_bytes_le(ptr.cast::<[u8; 32]>().as_ref()), f)
 }
